@@ -1,20 +1,20 @@
 import actionTypes from '../constants/actionTypes';
-import runtimeEnv from '@mars/heroku-js-runtime-env';
+import runtimeEnv from '@mars/heroku-js-runtime-env'
 
-function userLoggedIn(username){
+function userLoggedIn(username) {
     return {
         type: actionTypes.USER_LOGGEDIN,
         username: username
     }
 }
 
-function logout(){
+function logout() {
     return {
         type: actionTypes.USER_LOGOUT
     }
 }
 
-export function submitLogin(data){
+export function submitLogin(data) {
     const env = runtimeEnv();
     return dispatch => {
         return fetch(`${env.REACT_APP_API_URL}/signin`, {
@@ -24,24 +24,22 @@ export function submitLogin(data){
                 'Content-Type': 'application/json'
             },
             body: JSON.stringify(data),
-            mode: 'cors'})
-            .then( (response) => {
-                if (!response.ok) {
-                    throw Error(response.statusText);
-                }
-                return response.json();
-            })
-            .then( (res) => {
-                localStorage.setItem('username', data.username);
-                localStorage.setItem('token', res.token);
+            mode: 'cors'
+        }).then((response) => {
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
+            return response.json()
+        }).then((res) => {
+            localStorage.setItem('username', data.username);
+            localStorage.setItem('token', res.token);
 
-                dispatch(userLoggedIn(data.username));
-            })
-            .catch( (e) => console.log(e) );
+            dispatch(userLoggedIn(data.username));
+        }).catch((e) => console.log(e));
     }
 }
 
-export function submitRegister(data){
+export function submitRegister(data) {
     const env = runtimeEnv();
     return dispatch => {
         return fetch(`${env.REACT_APP_API_URL}/signup`, {
@@ -51,18 +49,15 @@ export function submitRegister(data){
                 'Content-Type': 'application/json'
             },
             body: JSON.stringify(data),
-            mode: 'cors'})
-            .then( (response) => {
-                if (!response.ok) {
-                    throw Error(response.statusText);
-                }
-                return response.json();
-            })
-            .then( (res) => {
-
-                dispatch(submitLogin(data));
-            })
-            .catch( (e) => console.log(e) );
+            mode: 'cors'
+        }).then((response) => {
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
+            return response.json()
+        }).then((res) => {
+            dispatch(submitLogin(data));
+        }).catch((e) => console.log(e));
     }
 }
 
@@ -70,6 +65,6 @@ export function logoutUser() {
     return dispatch => {
         localStorage.removeItem('username');
         localStorage.removeItem('token');
-        dispatch(logout());
+        dispatch(logout())
     }
 }
