@@ -1,5 +1,6 @@
 import actionTypes from '../constants/actionTypes';
-import runtimeEnv from '@mars/heroku-js-runtime-env';
+import runtimeEnv from '@mars/heroku-js-runtime-env'
+
 
 function moviesFetched(movies) {
     return {
@@ -24,7 +25,29 @@ function movieSet(movie) {
 
 export function setMovie(movie) {
     return dispatch => {
-        dispatch(movieSet(movie))
+        dispatch(movieSet(movie));
+    }
+}
+
+export function fetchMovie(movieId) {
+    const env = runtimeEnv();
+    return dispatch => {
+        return fetch(`${env.REACT_APP_API_URL}/movies/${movieId}?reviews=true`, {
+            method: 'GET',
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json',
+                'Authorization': localStorage.getItem('token')
+            },
+            mode: 'cors'
+        }).then((response) => {
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
+            return response.json()
+        }).then((res) => {
+            dispatch(movieFetched(res));
+        }).catch((e) => console.log(e));
     }
 }
 
@@ -38,40 +61,14 @@ export function fetchMovies() {
                 'Content-Type': 'application/json',
                 'Authorization': localStorage.getItem('token')
             },
-            mode: 'cors'})
-            .then( (response) => {
-                if (!response.ok) {
-                    throw Error(response.statusText);
-                }
-                return response.json();
-            })
-            .then( (res) => {
-                dispatch(moviesFetched(res));
-            })
-            .catch( (e) => console.log(e) );
-    }
-}
-
-export function fetchMovie(movieId){
-    const env = runtimeEnv();
-    return dispatch => {
-        return fetch(`${env.REACT_APP_API_URL}/movies/${movieId}?reviews=true`, {
-            method: 'GET',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json',
-                'Authorization': localStorage.getItem('token')
-            },
-            mode: 'cors'})
-            .then( (response) => {
-                if (!response.ok) {
-                    throw Error(response.statusText);
-                }
-                return response.json();
-            })
-            .then( (res) => {
-                dispatch(movieFetched(res));
-            })
-            .catch( (e) => console.log(e) );
+            mode: 'cors'
+        }).then((response) => {
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
+            return response.json()
+        }).then((res) => {
+            dispatch(moviesFetched(res));
+        }).catch((e) => console.log(e));
     }
 }
