@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { submitLogin } from '../actions/authActions';
 import { connect } from 'react-redux';
-import { Col, Form, FormGroup, FormControl, ControlLabel, Button } from 'react-bootstrap';
+import { Form, Button } from 'react-bootstrap';
 
 class Login extends Component {
 
@@ -34,30 +34,17 @@ class Login extends Component {
 
     render(){
         return (
-            <Form horizontal>
-                <FormGroup controlId="username">
-                    <Col componentClass={ControlLabel} sm={2}>
-                        Email
-                    </Col>
-                    <Col sm={10}>
-                        <FormControl onChange={this.updateDetails} value={this.state.details.username} type="email" placeholder="Email" />
-                    </Col>
-                </FormGroup>
+            <Form className='form-horizontal'>
+                <Form.Group controlId="username">
+                    <Form.Label>Email</Form.Label>
+                    <Form.Control onChange={this.updateDetails} value={this.state.details.username} type="email" placeholder="Enter email" />
+                </Form.Group>
 
-                <FormGroup controlId="password">
-                    <Col componentClass={ControlLabel} sm={2}>
-                        Password
-                    </Col>
-                    <Col sm={10}>
-                        <FormControl onChange={this.updateDetails} value={this.state.details.password} type="password" placeholder="Password" />
-                    </Col>
-                </FormGroup>
-
-                <FormGroup>
-                    <Col smOffset={2} sm={10}>
-                        <Button onClick={this.login}>Sign in</Button>
-                    </Col>
-                </FormGroup>
+                <Form.Group controlId="password">
+                    <Form.Label>Password</Form.Label>
+                    <Form.Control onChange={this.updateDetails} value={this.state.details.password}  type="password" placeholder="Password" />
+                </Form.Group>
+                <Button onClick={this.login}>Sign in</Button>
             </Form>
         )
     }
