@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import { fetchMovies } from '../actions/movieActions';
-import { setMovie } from '../actions/movieActions';
-import {connect} from "react-redux";
-import { Image } from 'react-bootstrap'
-import { Carousel } from 'react-bootstrap'
-import { Glyphicon } from 'react-bootstrap'
+import { fetchMovies } from "../actions/movieActions";
+import { setMovie } from "../actions/movieActions";
+import {connect} from 'react-redux';
+import {Image, Nav} from 'react-bootstrap';
+import { Carousel } from 'react-bootstrap';
+import { BsStarFill} from 'react-icons/bs'
 import {LinkContainer} from 'react-router-bootstrap';
 
 class MovieList extends Component {
@@ -29,9 +29,9 @@ class MovieList extends Component {
     }
 
     render() {
-        const MovieListCarousel= ({movieList}) => {
-            if (!movieList) { // evaluates to true if currentMovie is null
-                return <div>Loading...</div>;
+        const MovieListCarousel = ({movieList}) => {
+            if (!movieList) {
+                return <div>Loading....</div>
             }
 
             return (
@@ -40,20 +40,23 @@ class MovieList extends Component {
                         <Carousel.Item key={movie._id}>
                             <div>
                                 <LinkContainer to={'/movie/'+movie._id} onClick={()=>this.handleClick(movie)}>
-                                    <Image className="image" src={movie.imageUrl} thumbnail />
+                                    <Nav.Link><Image className="image" src={movie.imageUrl} thumbnail /></Nav.Link>
                                 </LinkContainer>
                             </div>
                             <Carousel.Caption>
                                 <h3>{movie.title}</h3>
-                                <Glyphicon glyph={'star'} /> {movie.avgRating} &nbsp;&nbsp; {movie.releaseDate}
+                                <BsStarFill glyph={'star'} /> {movie.avgRating} &nbsp;&nbsp; {movie.releaseDate}
                             </Carousel.Caption>
-                        </Carousel.Item>)}
-                </Carousel>)
+                        </Carousel.Item>
+                    )}
+
+                </Carousel>
+            )
         }
 
         return (
-          <MovieListCarousel movieList={this.props.movies} />
-        );
+            <MovieListCarousel movieList={this.props.movies} />
+        )
     }
 }
 
@@ -64,3 +67,4 @@ const mapStateToProps = state => {
 }
 
 export default connect(mapStateToProps)(MovieList);
+
